+++ trunk/s3backer.h
@@ -83,7 +83,7 @@ typedef uint32_t    s3b_block_t;
 /*
  * How many hex digits we will use to print a block number.
  */
-#define S3B_BLOCK_NUM_DIGITS    (sizeof(s3b_block_t) * 2)
+#define S3B_BLOCK_NUM_DIGITS    ((int)(sizeof(s3b_block_t) * 2))
 
 /* Logging function type */
 typedef void        log_func_t(int level, const char *fmt, ...) __attribute__ ((__format__ (__printf__, 2, 3)));
